ads-sim/requirements.txt
letta
praw
tqdm
requests
python-dotenv
uvicorn
fastapi